Use one request to fetch all patern code editors
